routes authorization

+++ src/routes/memories.ts
@@ -3,8 +3,16 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
+
       orderBy: {
         createdAt: "asc",
       },
@@ -19,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/:id", async (request) => {
+  app.get("/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -31,6 +39,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     return {
       memory,
@@ -51,7 +63,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: request.id,
+        userId: request.user.sub,
       },
     });
 
@@ -60,7 +72,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     };
   });
 
-  app.put("/:id", async (request) => {
+  app.put("/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -75,7 +87,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -91,12 +113,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     };
   });
 
-  app.delete("/:id", async (request) => {
+  app.delete("/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
